Add more documentation about short vs. long-form DIDs
It took me a while to piece together the differences, relationships
between and implications of Sidetree, ION, and DID Resolution.

This commit hopefully will help out future developers that are trying to
break into this domain.

Of note, this commit results in the IG taking a stronger stance on using
long-form DIDs over short-form. Happy to change that if desirable!