docs/index.md: Add Golang resource for learning
I wrote some libraries that can be composed together for generate smart health card QR codes for COVID-19 immunizations. The code is in Golang, and might be a nice addition to the JavaScript/Node and Python/Jupyter examples repos listed here.

My repo is here: https://github.com/amitkgupta/go-smarthealthcards

If you like it, please consider adding it to your docs! I wasn't able to find Go-based solutions here or elsewhere on the web, so I wrote one myself and I hope it can be of use to others!