Add Pipenv for local build support
Adding the redirection plugin makes it a bit harder to get this set
up locally, so add Pipenv files for dependency management and update
README accordingly.

Note that the redirect plugin is bundled in the Docker image used
in the GitHub Action, so no change is needed there unless more
plugins are added in the future.