Add suggested Deep Link format
The suggested format uses the URL fragment (delimited with `#`) and starts with an opening parenthesis `(`, followed by the raw JWS, a comma separating any further JWS if necessary, and a closing parenthesis `)`. This allows for percent-encoding free representation of the payload(s) in the fragment and allows detection of the payload type by combining knowledge of the path (suggested as containing `SMARTHealthCard` and the `()` bounds.